// src/generated/ApplicationUpdateProgressDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApplicationUpdateOperationDto } from "./ApplicationUpdateOperationDto";

export type ApplicationUpdateProgressDto = { operation: ApplicationUpdateOperationDto, downloadedBytes: number | null, totalBytes: number | null, };